\name{flip}
\alias{flip}
\title{Flip Genotypes}
\usage{
  flip(SNP, sep = "", outSep = sep)
}
\arguments{
  \item{SNP}{A vector of genotypes for a particular locus.}

  \item{sep}{The separator between each allele.}

  \item{outSep}{The output separator to use.}
}
\description{
  Given a set of genotypes, flip them.
}
\keyword{internal}

