% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_snp_api.R
\name{get_frequency}
\alias{get_frequency}
\title{Internal function to get the frequency of the variants from
different studies.}
\usage{
get_frequency(Class, primary_info, study = "GnomAD")
}
\arguments{
\item{Class}{What kind of variant is the rsid. Accepted options are "snv", "snp" and "delins".}

\item{primary_info}{refsnp entry read in JSON format}

\item{study}{Study from which frequency information is obtained. Possibilities
include: GnomAD (default), 1000Genomes, ALSPAC, Estonian, NorthernSweden, TWINSUK}
}
\description{
Internal function to get the frequency of the variants from
different studies.
}
