% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_snp_api.R
\name{get_gene_names}
\alias{get_gene_names}
\title{Internal function to get gene names.}
\usage{
get_gene_names(primary_info)
}
\arguments{
\item{primary_info}{refsnp entry read in JSON format}
}
\description{
If multiple gene names are encountered they are collapsed with a
"/".
}
