% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_snp_api.R
\name{get_placements}
\alias{get_placements}
\title{Internal function to get the position, alleles, assembly, hgvs notation}
\usage{
get_placements(primary_info)
}
\arguments{
\item{primary_info}{refsnp entry read in JSON format}
}
\description{
Internal function to get the position, alleles, assembly, hgvs notation
}
