% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSearch.R
\name{ld_search}
\alias{ld_search}
\title{This function is defunct.}
\usage{
ld_search(...)
}
\value{
a error message telling the user that this function has been removed
}
\description{
This function is defunct.
}
\keyword{internal}
