% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-defunct.R
\name{LDSearch}
\alias{LDSearch}
\title{This function is defunct.}
\usage{
LDSearch(...)
}
\value{
a error message telling the user that this function has been removed
}
\description{
This function is defunct.
}
\keyword{internal}
