% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_snp_query.R
\name{NCBI_snp_query}
\alias{NCBI_snp_query}
\alias{ncbi_snp_query}
\title{Query NCBI's dbSNP for information on a set of SNPs}
\usage{
NCBI_snp_query(SNPs, ...)

ncbi_snp_query(SNPs, ...)
}
\arguments{
\item{SNPs}{A vector of SNPs (rs numbers).}

\item{...}{Further named parameters passed on to \code{\link[httr]{config}} 
to debug curl. See examples.}
}
\value{
A dataframe with columns:
\itemize{
\item \code{Query:} The rs ID that was queried.
\item \code{Chromosome:} The chromosome that the marker lies on.
\item \code{Marker:} The name of the marker. If the rs ID queried
has been merged, the up-to-date name of the marker is returned here, and
a warning is issued.
\item \code{Class:} The marker's 'class'. See
\url{http://www.ncbi.nlm.nih.gov/projects/SNP/snp_legend.cgi?legend=snpClass}
for more details.
\item \code{Gene:} If the marker lies within a gene (either within the exon
or introns of a gene), the name of that gene is returned here; otherwise,
\code{NA}. Note that
the gene may not be returned if the marker lies too far upstream or downstream
of the particular gene of interest.
\item \code{Alleles:} The alleles associated with the SNP if it is a
SNV; otherwise, if it is an INDEL, microsatellite, or other kind of
polymorphism the relevant information will be available here.
\item \code{Major:} The major allele of the SNP, on the forward strand,
given it is an SNV; otherwise, \code{NA}.
\item \code{Minor:} The minor allele of the SNP, on the forward strand,
given it is an SNV; otherwise, \code{NA}.
\item \code{MAF:} The minor allele frequency of the SNP, given it is an SNV.
This is drawn from the current global reference population used by NCBI.
\item \code{BP:} The chromosomal position, in base pairs, of the marker,
as aligned with the current genome used by dbSNP. we add 1 to the base 
pair position in the BP column in the output data.frame to agree with 
what the dBSNP website has.
}
}
\description{
This function queries NCBI's dbSNP for information related to the latest
dbSNP build and latest reference genome for information on the vector
of SNPs submitted.
}
\details{
Note that you are limited in the number of SNPs you pass in to one 
request because URLs can only be so long. Around 600 is likely the max you 
can pass in, though may be somewhat more. Break up your vector of SNP 
codes into pieces of 600 or less and do repeated requests to get all data.
}
\note{
\code{ncbi_snp_query} is a synonym of \code{NCBI_snp_query} - we'll 
make \code{NCBI_snp_query} defunct in the next version
}
\examples{
\dontrun{
## an example with both merged SNPs, non-SNV SNPs, regular SNPs,
## SNPs not found, microsatellite
SNPs <- c("rs332", "rs420358", "rs1837253", "rs1209415715", "rs111068718")
ncbi_snp_query(SNPs)
# ncbi_snp_query("123456") ##invalid: must prefix with 'rs'
ncbi_snp_query("rs420358")
ncbi_snp_query("rs332") # warning that its merged into another, try that
ncbi_snp_query("rs121909001")
ncbi_snp_query("rs1837253")
# warning that no data available, returns 0 length data.frame
ncbi_snp_query("rs1209415715")
# warning that chromosomal information may be unmapped 
ncbi_snp_query("rs111068718") 

ncbi_snp_query(SNPs='rs9970807')$BP

# Curl debugging
ncbi_snp_query("rs121909001")
library("httr")
ncbi_snp_query("rs121909001", config=verbose())
snps <- c("rs332", "rs420358", "rs1837253", "rs1209415715", "rs111068718")
ncbi_snp_query(snps, config=progress())
}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/projects/SNP/}
}
\seealso{
\code{\link{ncbi_snp_query2}}
}

