% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-defunct.R
\name{NCBI_snp_query2}
\alias{NCBI_snp_query2}
\title{This function is defunct.}
\usage{
NCBI_snp_query2(...)
}
\value{
a error message telling the user that this function has been deprecated in favour of another function
}
\description{
This function is defunct.
}
\keyword{internal}
