% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_snp_par.R
\name{NCBI_snp_query_par}
\alias{NCBI_snp_query_par}
\title{Query NCBI's dbSNP for information on a set of SNPs}
\usage{
NCBI_snp_query_par(SNPs, blocksize = 200, numcores = 36, hownice = "very")
}
\arguments{
\item{SNPs}{A vector of SNPs (rs numbers).}

\item{blocksize}{how many SNPs to look at each time}

\item{numcores}{cores for parallel}

\item{hownice}{how nice? options are: c("very", "somewhat", "not")}
}
\description{
Query NCBI's dbSNP for information on a set of SNPs
}

