% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-defunct.R
\name{rsnps-defunct}
\alias{rsnps-defunct}
\title{Defunct functions in rsnps}
\description{
- `LDSearch()`: Function name changed to [ld_search]
- `NCBI_snp_query()`: Function name changed to [ncbi_snp_query]
- `NCBI_snp_query2()`: Function name changed to [ncbi_snp_query2]
}
