% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap}
\alias{swap}
\title{Swap Elements in a Vector}
\usage{
swap(vec, from, to = names(from), ...)
}
\arguments{
\item{vec}{A character vector, or vector coercible to character.}

\item{from}{A vector of elements to map from.}

\item{to}{A vector of elements to map to.}

\item{...}{Optional arguments passed to \code{\link[=match]{match()}}}
}
\description{
Swap Elements in a Vector
}
\keyword{internal}
