% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgensnp.R
\name{allgensnp}
\alias{allgensnp}
\title{Get openSNP genotype data for all users at a particular snp.}
\usage{
allgensnp(snp = NA, ...)
}
\arguments{
\item{snp}{(character) A SNP name}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
data.frame of genotypes for all users at a certain SNP
}
\description{
Get openSNP genotype data for all users at a particular snp.
}
\examples{
\dontrun{
x <- allgensnp(snp = 'rs7412')
head(x)
}
}
\seealso{
Other opensnp-fxns: \code{\link{allphenotypes}},
  \code{\link{annotations}}, \code{\link{download_users}},
  \code{\link{fetch_genotypes}}, \code{\link{genotypes}},
  \code{\link{phenotypes_byid}}, \code{\link{phenotypes}},
  \code{\link{users}}
}
\concept{opensnp-fxns}
