% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations}
\alias{annotations}
\title{Get all openSNP phenotypes, their variations, and how many users have data
available for a given phenotype.}
\usage{
annotations(
  snp = NA,
  output = c("all", "plos", "mendeley", "snpedia", "metadata"),
  ...
)
}
\arguments{
\item{snp}{SNP name.}

\item{output}{Name the source or sources you want annotations from (options
are: 'plos', 'mendeley', 'snpedia', 'metadata'). 'metadata' gives the
metadata for the response.}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
data.frame of results
}
\description{
Either return data.frame with all results, or output a list, then call
the characteristic by id (parameter = "id") or
name (parameter = "characteristic").
}
\examples{
\dontrun{
# Get all data
## get just the metadata
annotations(snp = "rs7903146", output = "metadata")

## just from plos
annotations(snp = "rs7903146", output = "plos")

## just from snpedia
annotations(snp = "rs7903146", output = "snpedia")

## get all annotations
annotations(snp = "rs7903146", output = "all")
}
}
\seealso{
Other opensnp-fxns: 
\code{\link{allgensnp}()},
\code{\link{allphenotypes}()},
\code{\link{download_users}()},
\code{\link{fetch_genotypes}()},
\code{\link{genotypes}()},
\code{\link{phenotypes_byid}()},
\code{\link{phenotypes}()},
\code{\link{users}()}
}
\concept{opensnp-fxns}
