% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-defunct.R
\name{rsnps-defunct}
\alias{rsnps-defunct}
\title{Defunct functions in rsnps}
\description{
\itemize{
\item \code{LDSearch()}: Function name changed to \link{ld_search}
\item \code{ld_search()}: The Broad Institute took the service down, see
https://www.broadinstitute.org/snap/snap
\item \code{NCBI_snp_query()}: Function name changed to \link{ncbi_snp_query}
\item \code{NCBI_snp_query2()}: Function name changed to \link{ncbi_snp_query}
\item \code{ncbi_snp_summary()}: Function name changed to \link{ncbi_snp_query}
\item \code{ncbi_snp_query2()}: Function name changed to \link{ncbi_snp_query}
}
}
