% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnps-package.R
\docType{package}
\name{rsnps-package}
\alias{rsnps}
\alias{rsnps-package}
\title{Get SNP (Single-Nucleotide Polymorphism) Data on the Web}
\description{
This package gives you access to data from OpenSNP (https://opensnp.org)
via their API (https://opensnp.org/faq#api).
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Kevin Ushey \email{kevinushey@gmail.com}

Hao Zhu \email{haozhu233@gmail.com}
}

