% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_users.R
\docType{data}
\name{rsnpsCache}
\alias{rsnpsCache}
\title{rsnps environment}
\format{\preformatted{<environment: 0x1037cc188> 
}}
\usage{
rsnpsCache
}
\description{
rsnps environment
}
\keyword{internal}

