% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_users.R
\docType{data}
\name{rsnpsCache}
\alias{rsnpsCache}
\title{rsnps environment}
\format{An object of class \code{environment} of length 0.}
\usage{
rsnpsCache
}
\description{
rsnps environment
}
\keyword{internal}

