\name{swap}
\alias{swap}
\title{Swap Elements in a Vector}
\usage{
swap(vec, from, to = names(from), ...)
}
\arguments{
  \item{vec}{A character vector, or vector coercable to
  character.}

  \item{from}{A vector of elements to map from.}

  \item{to}{A vector of elements to map to.}

  \item{...}{Optional arguments passed to
  \code{\link{match}}.}
}
\description{
Swap Elements in a Vector
}
\keyword{internal}

